example tests: do not exclude Darwin
This check seems suspicious to me to begin with.  It is surely
here to reject Windows, by matching MSWin32 and cygwin.  Even that
seems overly exclusive to me.  Can this not be made to work there?

But the real problem is that "darwin" is now a popular value for
the unix macOS.