Test::Unit::Result: cope with non-Error.pm-based exceptions
Imagine this code in a test case:

    my sub contrived_example {
        die [ "array-based exception" ];
    }

    sub test_something ($self) {
        contrived_example();
        $self->assert(0); # always fail
    }

No matter what, one would expect this test to be either a failure
(there's a false assertion) or an error (a non-Failure exception is
raised).  Instead, **this test would pass** before this commit.

The issue is that the code (in Result.pm) that runs a test is inside a
try/catch structure that only had handlers for exceptions that are isa
Test::Unit::Failure or Error.  Any other exception was simply dropped on
the floor, because there was no "otherwise" clause for handling
yet-unhandled exceptions.

This commit adds one.  That clause will convert any such exception into
a Test::Unit::Error, first stringifying the original exception.  The
exception is stringified rather than included because the
Test::Unit::Error is then serialized and sent over the IPC pipe.  Some
exceptions (in our case, JMAP::Tester::Abort objects) include code
references that can't be serialized.